Clearly separating the tornado version from the base. It prevents unnecessary dependency on Tornado when using the Gevent version.
